requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11
django-filter==1.0.2
djangorestframework==3.6.2
gunicorn==19.7.1
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
requests==2.18.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
whitenoise==3.3.0
